server/requirements.txt
alembic==1.4.2
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
psycopg2==2.8.5
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.18
sqlalchemy-mixins==1.2.1
SQLAlchemy-Utils==0.36.6
Werkzeug==1.0.1
